OT231-25 : Move and Rename Security Authentication and AuthenticationResponse files
