+++ src/main/java/com/alkemy/ong/infrastructure/config/spring/security/SecurityConfig.java
@@ -56,7 +56,7 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
 
   @Override
   protected void configure(HttpSecurity http) throws Exception {
-  
+
     http.authorizeRequests()
             .antMatchers("/auth/login")
             .permitAll()
